Refactor: 검색창 분리 및 서버 컴포넌트화
