+++ src/components/common/Header/Header.module.scss
@@ -20,35 +20,6 @@
   grid-area: logo;
 }
 
-.searchBarWrapper {
-  position: relative;
-  grid-area: searchBar;
-  width: 45rem;
-  height: 4rem;
-}
-
-.searchButton {
-  position: absolute;
-  top: 1rem;
-  left: 1rem;
-}
-
-.searchBar {
-  width: 100%;
-  height: 100%;
-  padding: 1rem;
-  padding-left: 4.2rem;
-  font-size: 1.4rem;
-  line-height: 2.2rem;
-  color: $gray-500;
-  background-color: $gray-200;
-  border: 0;
-  border-radius: 1rem;
-  outline: none;
-
-  @include font-normal;
-}
-
 @mixin buttonWrapper {
   display: flex;
   grid-area: buttonWrapper;
@@ -107,10 +78,6 @@
     padding: 1.5rem 3rem;
   }
 
-  .searchBarWrapper {
-    width: 34.4rem;
-  }
-
   .buttonWrapper {
     gap: 1.2rem;
   }
@@ -126,11 +93,6 @@
     width: 100%;
     min-width: none;
   }
-
-  .searchBarWrapper {
-    width: 100%;
-  }
-
   .logo {
     width: 8.4rem;
     height: 3rem;
