+++ src/components/common/Header/Header.tsx
@@ -1,25 +1,16 @@
-'use client';
-
 import Image from 'next/image';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { useState, useEffect } from 'react';
-import { SEARCH_ICON, HEADER_LOGO } from '@/utils/constants';
-import Cookies from 'js-cookie';
+import { HEADER_LOGO } from '@/utils/constants';
+import { cookies } from 'next/headers';
 import Notification from './notification';
 import styles from './Header.module.scss';
 import { ButtonListType } from './types';
+import SearchBox from './SearchBox';
+import LogoutButton from './LogoutButton';
 
 export default function Header() {
-  const router = useRouter();
-  const [memberType, setMemberType] = useState('none');
-  const [input, setInput] = useState('');
-  const logout = () => {
-    const removeData = ['id', 'token', 'type', 'shopId', 'userId'];
-    removeData.forEach((data) => Cookies.remove(data));
-    setMemberType('none');
-    router.push('/');
-  };
+  const cookieStore = cookies();
+  const memberType = cookieStore.get('type')?.value ?? 'none';
 
   const BUTTON_LIST: ButtonListType = {
     none: {
@@ -31,60 +22,32 @@ export default function Header() {
     },
     employer: {
       buttonList: [
-        { name: '내 가게', href: '/shopinfo' },
-        { name: '로그아웃', href: '', onClick: logout },
+        { name: '내 가게', href: '/mystore' },
+        { name: '로그아웃', href: '' },
       ],
       notification: true,
     },
     employee: {
       buttonList: [
-        { name: '내 프로필', href: '/profile' },
-        { name: '로그아웃', href: '', onClick: logout },
+        { name: '내 프로필', href: '/myprofile' },
+        { name: '로그아웃', href: '' },
       ],
       notification: true,
     },
   };
   const buttonType = BUTTON_LIST[memberType];
 
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setInput(event.target.value);
-  };
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === 'Enter' && input) {
-      router.push(`/?keyword=${input}`);
-    }
-  };
-  useEffect(() => {
-    const type = Cookies.get('type');
-    if (type === 'employer' || type === 'employee') {
-      setMemberType(type);
-      return;
-    }
-    setMemberType('none');
-  }, []);
-
   return (
     <div className={styles.wrapper}>
       <div className={styles.contentWrapper}>
-        <Link href="/">
+        <a href="/">
           <Image src={HEADER_LOGO} alt="logo" width={112} height={40} className={styles.logo} priority />
-        </Link>
-        <form className={styles.searchBarWrapper} onSubmit={(e) => e.preventDefault()}>
-          <Image src={SEARCH_ICON} alt="search" width={20} height={20} className={styles.searchButton} priority />
-          <input
-            className={styles.searchBar}
-            value={input}
-            onChange={handleInputChange}
-            onKeyDown={handleKeyDown}
-            placeholder="가게 이름으로 찾아보세요"
-          />
-        </form>
+        </a>
+        <SearchBox />
         <div className={memberType === 'none' ? styles.buttonWrapperSmall : styles.buttonWrapper}>
           {buttonType.buttonList.map((button) =>
-            button.onClick ? (
-              <div className={styles.button} key={button.name} onClick={button.onClick} role="presentation">
-                {button.name}
-              </div>
+            button.name === '로그아웃' ? (
+              <LogoutButton key="로그아웃" />
             ) : (
               <Link href={button.href} key={button.name}>
                 <div className={styles.button}>{button.name}</div>
