+++ src/components/common/Header/SearchBox.module.scss
@@ -1,0 +1,42 @@
+@import '@/styles/_index';
+@import '@/styles/_media';
+
+.searchBarWrapper {
+  grid-area: searchBar;
+  width: 45rem;
+  height: 4rem;
+  position: relative;
+}
+
+.searchButton {
+  position: absolute;
+  left: 1rem;
+  top: 1rem;
+}
+
+.searchBar {
+  width: 100%;
+  height: 100%;
+  border-radius: 1rem;
+  padding: 1rem;
+  padding-left: 4.2rem;
+  background-color: $gray-200;
+  color: $gray-500;
+  outline: none;
+  border: 0;
+  @include font-normal;
+  font-size: 1.4rem;
+  line-height: 2.2rem;
+}
+
+@include media(tablet) {
+  .searchBarWrapper {
+    width: 34.4rem;
+  }
+}
+
+@include media(mobile) {
+  .searchBarWrapper {
+    width: 100%;
+  }
+}
